Refactor insights page into modular components
