+++ src/mock/insightsChartData.ts
@@ -1,0 +1,27 @@
+export const acplData = [
+  { phase: 'Opening', acpl: 32 },
+  { phase: 'Middle', acpl: 47 },
+  { phase: 'Late', acpl: 38 },
+  { phase: 'Endgame', acpl: 28 },
+];
+
+export const ecoData = [
+  { eco: 'B20', score: 70 },
+  { eco: 'C50', score: 60 },
+  { eco: 'D30', score: 65 },
+  { eco: 'E60', score: 55 },
+];
+
+export const lossReasonsData = [
+  { reason: 'Pawn push', games: 5 },
+  { reason: 'Knight fork', games: 3 },
+  { reason: 'Hung mate', games: 2 },
+  { reason: 'Impulsive', games: 8 },
+];
+
+export const timeControlData = [
+  { control: 'Bullet', impulsive: 5, slow: 1 },
+  { control: 'Blitz', impulsive: 7, slow: 2 },
+  { control: 'Rapid', impulsive: 4, slow: 3 },
+  { control: 'Classical', impulsive: 1, slow: 4 },
+];
